(WIP) Rework robot representation
