--- phobos/smurf/__init__.py
@@ -1,4 +1,0 @@
-from .robot import Smurf
-from .motors import Motor
-from .poses import JointPositionSet, JointPosition
-from .hyrodyn import Submechanism, Exoskeleton
